requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
pydantic==2.7.1
openai==1.76.0
openai-agents==0.0.13
psycopg2-binary==2.9.9
python-dotenv==1.0.1
